refactor(key-server): pull out ca_manager funcs
All the free functions in here don't really need to be in ca_manager
because a lot will be reused in the tls certificate side of things, so
split them out and split out their tests too.
